Case Designer page layout
